+++ src/main/java/graphics/kiln/bakedminecraftmodels/data/InstanceBatch.java
@@ -6,8 +6,18 @@
 
 package graphics.kiln.bakedminecraftmodels.data;
 
+import graphics.kiln.bakedminecraftmodels.gl.GlSsboRenderDispacher;
+import graphics.kiln.bakedminecraftmodels.mixin.renderlayer.MultiPhaseParametersAccessor;
+import graphics.kiln.bakedminecraftmodels.mixin.renderlayer.MultiPhaseRenderPassAccessor;
+import graphics.kiln.bakedminecraftmodels.model.VboBackedModel;
 import graphics.kiln.bakedminecraftmodels.ssbo.SectionedPersistentBuffer;
+import it.unimi.dsi.fastutil.ints.IntArrays;
+import net.minecraft.client.render.LightmapTextureManager;
+import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.VertexFormat;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.util.math.MathHelper;
+import net.minecraft.util.math.Matrix4f;
 import org.lwjgl.system.MemoryUtil;
 
 import java.util.ArrayList;
@@ -15,7 +25,8 @@ import java.util.List;
 
 public class InstanceBatch {
 
-    private final List<BakingData.PerInstanceData> instances;
+    private final SectionedPersistentBuffer partBuffer;
+    private final List<PerInstanceData> instances;
     private final MatrixEntryList matrices;
 
     private int[] primitiveIndices;
@@ -24,9 +35,10 @@ public class InstanceBatch {
     private VertexFormat.IntType indexType;
     private long indexOffset;
 
-    public InstanceBatch(int initialSize) {
+    public InstanceBatch(int initialSize, SectionedPersistentBuffer partBuffer) {
         this.instances = new ArrayList<>(initialSize);
         this.matrices = new MatrixEntryList(initialSize);
+        this.partBuffer = partBuffer;
     }
 
     public void reset() {
@@ -48,13 +60,102 @@ public class InstanceBatch {
     }
 
     public void writeInstancesToBuffer(SectionedPersistentBuffer buffer) {
-        for (BakingData.PerInstanceData perInstanceData : instances) {
+        for (PerInstanceData perInstanceData : instances) {
             perInstanceData.writeToBuffer(buffer);
         }
     }
 
-    public void addInstance(BakingData.PerInstanceData instanceData) {
-        instances.add(instanceData);
+    public void addInstance(VboBackedModel model, RenderLayer renderLayer, MatrixStack.Entry baseMatrixEntry, float red, float green, float blue, float alpha, int overlay, int light) {
+        int overlayX = overlay & 0xFFFF;
+        int overlayY = overlay >> 16 & 0xFFFF;
+        int lightX = light & (LightmapTextureManager.MAX_BLOCK_LIGHT_COORDINATE | 0xFF0F);
+        int lightY = light >> 16 & (LightmapTextureManager.MAX_BLOCK_LIGHT_COORDINATE | 0xFF0F);
+
+        MultiPhaseParametersAccessor multiPhaseParameters = (MultiPhaseParametersAccessor) (Object) ((MultiPhaseRenderPassAccessor) renderLayer).getPhases();
+
+        // this can happen if the model didn't render any modelparts,
+        // in which case it makes sense to not try to render it anyway.
+        if (matrices.isEmpty()) return;
+        long partIndex = matrices.writeToBuffer(partBuffer, baseMatrixEntry);
+
+        // While we have the matrices around, do the transparency processing if required
+        if (GlSsboRenderDispacher.requiresIndexing(multiPhaseParameters)) {
+            // Build the camera transforms from all the part transforms
+            // This is how we quickly measure the depth of each primitive - find the
+            // camera's position in model space, rather than applying the matrix multiply
+            // to the primitive's position.
+            float[] cameraPositions = new float[matrices.getLargestPartId() * 3];
+            for (int partId = 0; partId < cameraPositions.length; partId++) {
+                Matrix4f m;
+                if (!matrices.getElementWritten(partId)) {
+                    m = baseMatrixEntry.getModel();
+                } else {
+                    MatrixStack.Entry entry = matrices.get(partId);
+                    if (entry != null) {
+                        m = entry.getModel();
+                    } else {
+                        // skip empty part
+                        continue;
+                    }
+                }
+
+                // The translation of the inverse of a transform matrix is the negation of
+                // the transposed rotation times the transform of the original matrix.
+                //
+                // The above only works if there's no scaling though - to correct for that, we
+                // can find the length of each column is the scaling factor for x, y or z depending
+                // on the column number. We then divide each of the output components by the
+                // square of the scaling factor - since we're multiplying the scaling factor in a
+                // second time with the matrix multiply, we have to divide twice (same as divide by sq root)
+                // to get the actual inverse.
+
+                // Using fastInverseSqrt might be playing with fire here
+                float undoScaleX = 1 / MathHelper.sqrt(m.a00 * m.a00 + m.a10 * m.a10 + m.a20 * m.a20);
+                float undoScaleY = 1 / MathHelper.sqrt(m.a01 * m.a01 + m.a11 * m.a11 + m.a21 * m.a21);
+                float undoScaleZ = 1 / MathHelper.sqrt(m.a02 * m.a02 + m.a12 * m.a12 + m.a22 * m.a22);
+
+                int arrayIdx = partId * 3;
+                cameraPositions[arrayIdx] = -(m.a00 * m.a03 + m.a10 * m.a13 + m.a20 * m.a23) * undoScaleX * undoScaleX;
+                cameraPositions[arrayIdx + 1] = -(m.a01 * m.a03 + m.a11 * m.a13 + m.a21 * m.a23) * undoScaleY * undoScaleY;
+                cameraPositions[arrayIdx + 2] = -(m.a02 * m.a03 + m.a12 * m.a13 + m.a22 * m.a23) * undoScaleZ * undoScaleZ;
+            }
+
+            float[] primitivePositions = model.getPrimitivePositions();
+            int[] primitivePartIds = model.getPrimitivePartIds();
+
+            float[] primitiveSqDistances = new float[primitivePositions.length];
+            int[] primitiveIndices = new int[primitivePositions.length];
+            for (int i = 0; i < primitivePositions.length; i++) {
+                primitiveIndices[i] = i;
+            }
+            int skippedPrimitives = 0;
+
+            for (int prim = 0; prim < primitivePositions.length; prim++) {
+                // skip if written as null
+                int partId = primitivePartIds[prim];
+                if (matrices.getElementWritten(partId) && matrices.get(partId) == null) {
+                    primitiveSqDistances[prim] = Float.MIN_VALUE;
+                    skippedPrimitives++;
+                }
+
+                int primPosIdx = prim * 3;
+                float x = primitivePositions[primPosIdx];
+                float y = primitivePositions[primPosIdx + 1];
+                float z = primitivePositions[primPosIdx + 2];
+
+                int camPosIdx = partId * 3;
+                float deltaX = x - cameraPositions[camPosIdx];
+                float deltaY = y - cameraPositions[camPosIdx + 1];
+                float deltaZ = z - cameraPositions[camPosIdx + 2];
+                primitiveSqDistances[prim] = deltaX * deltaX + deltaY * deltaY + deltaZ * deltaZ;
+            }
+
+            IntArrays.quickSort(primitiveIndices, (i1, i2) -> Float.compare(primitiveSqDistances[i1], primitiveSqDistances[i2]));
+
+            setPrimitiveIndices(primitiveIndices, skippedPrimitives);
+        }
+
+        instances.add(new PerInstanceData(partIndex, red, green, blue, alpha, overlayX, overlayY, lightX, lightY));
     }
 
     public int size() {
